We have an empty Dinosaur class. As proud of that empty class as I am, I think we need to start thinking about what we need this class to actually do - how we want it to behave! This totally depends on your app - and what you need to use each class for. But... hmm... let's see. I definitely want to be able to set the length on a Dinosaur - because maybe we need to render that somewhere. Oh, and if the length is not set, it should probably default to be 0.
Our First Example

Wait. Right there. Did you notice? I just described an example of how our Dinosaur class should work! All we need to do is translate that into a phpspec example! Create a new function, start it with it_ - because that's what phpspec requires... and also because that helps us create descriptive & readable method names. How about: it_should_default_to_zero_length()

Inside, remember: the goal is to pretend like we're inside the Dinosuar class. When each example is executed, phpspec will instantiate a Dinosaur object behind the scenes and we can reference it via $ this. That's total and absolute magic... and we'll get to find out exactly how it works a bit later.

Anyways, right now: just imagine that $ this is a Dinosaur object. To show an example of how the length should be 0 by default, we will literally write example code. For our app, I want to be able to call a getLength() method to get the length. Cool! Write: $ this->getLength() as if that method already existed.

And because we want our Dinosaur's length to default to 0, call a matcher function to assert that: ->shouldReturn(0).

Matchers!
How... weird, but cool is that?! At any point in phpspec, you can say ->should to call one of phpspec's many 'matcher' functions. These are equivalent to the assert() functions in PHPUnit - the difference is purely style. Instead of saying $ this->assertEquals(0, $ dinosuar->getLength()) like you would in PHPUnit - boring - you say $ this->getLength()->shouldReturn(0). The whole line reads like a clear English sentence!

And fortunately, even though phpspec is using some legit sourcery to make this all work, PhpStorm has great support for auto-completing these matcher functions. Oh, and, by the way, every matcher always starts with should or shouldNot. That's just a rule - and we'll learn why later.

Find your browser and go to https://phpspec.net. Click into the manual and go to the matchers section. Nice! phpspec has a huge number of matchers... and someone even thought to document them! Amazing! Right now, we're using one called the Identity Matcher, which allows you to use shouldBe(), shouldBeEqualTo() shouldReturn() or shouldEqual(). These are all different ways to compare values using ===.

There's also a Comparison Matcher where you can say shouldBeLike() to compare values using ==. And there are many, many, many more to geek out over. We'll learn the most useful ones along the way.