Run: ./vendor/bin/phpspec run
Generating the Missing Method
We now have one new example of how we want the Dinosaur class to work. Will this example already pass? Of course not! We don't even have a getLength() method yet! But, run phpspec anyways - and prepared to be... amazed:
./vendor/bin/phpspec run
Yes! I love failure! The getLength() method is not found. Oh, but check it out! Just like before, it realizes that we're describing some behavior that doesn't exist and asks us if we want it to do our job for us! Of course we do!

Let's go check it out! Not bad! It generates the method but, unless your version of phpspec has become self-aware - in which case... let me know what version you're using - it has no idea what to put inside the method.
And so, after phpspec generated the code, it automatically re-ran itself, but the new example still fails:
it should default to zero length, expected integer 0 but got null,

This is the phpspec flow!
One: describe behavior with an example.
Two: phpspec generates as much as it can.
Three: we fill in the logic. Four: profit!