Filling in the Logic
And... actually, the rules of TDD say that we should fill in the method with as little code as possible to get the test to pass - including just hardcoding a value if you can! But... more on that craziness later.

Right now, let's fill this in for real. So, hmm... because we know each Dinosaur can have a different length, we will probably need a $ length property. And, ah yes, it needs to default to 0 - that's something we decided during the 'spec' or 'description' process. Inside the method, return $ this->length. Oh, and to be super-cool, add the int return type. Viva return types!

Our class should now behave like our example expects. Let's see if phpspec agrees! Run it:
./vendor/bin/phpspec run
Woohoo! It passes!