Describing setLength()
But... nobody want's to visit a dinosaur park full of dinosaurs with zero length. We need a way to set the length. How do we want to do that? There's no right answer: it depends on your app. For example, we could make it a constructor argument. Or, it might make sense in your app to have, for example, some updateSpecs() method where you pass the length along with a few other things about your dinosaur. Or, you might need a simple setLength() method. The cool thing is that phpspec forces you to think about this - it forces you to ask: How will this class be used? Do I really need a setLength() method? Or will the user set a bunch of details all at once, and so a more descriptive updateSpecs() method is more clear?

Again... there's no right or wrong answer - sorry! For our app, let's create an example showing setLength(): function it_should_allow_to_set_length(). Inside, pretending that $ this is a Dinosaur object, let's show how this should work: call $ this->setLength() and pass it, how about 9. After that, we should be able to call $ this->getLength()->shouldReturn(9).

Done! Oh, and I want you to notice one cool thing: we now get autocomplete on the getLength() method! PhpStorm has great phpspec integration and knows that we're allowed to use  like a Dinosaur object. We don't have auto-completion for setLength(), because that method doesn't exist yet.