One of the main goals of a spec class is for it to communicate the behavior of our class through readable and natural language. More important than being a test, this class is meant to be documentation. If the function names or code inside the functions aren't naturally readable - you're at risk of angering the phpspec gods!
For example, saying $ this->getLength()->shouldReturn(9) does read like a normal, English sentence. But let's pretend for a minute that this language does not sound clear - maybe we're using a matcher that works, but just feels unnatural. In that case, we can invent our own language. Check it out: create a new example function: it_should_default_to_zero_length_using_custom_matcher().

Inside, let's show this same behavior, but in a different way - how about $ this->getLength()->shouldReturnZero().

That's great language! But, as you probably noticed, PhpStorm did not auto-complete that matcher function. That's because... I just made that up! There is no built-in matcher that allows us to say shouldReturnZero().

To prove it, run spec!
./vendor/bin/phpspec run
No returnZero matcher found. But, if this is the language that is most natural, we can and should make it work. How? By creating our own matcher.