Overriding getMatchers()
At the top of your class... or really anywhere, go to the Code -> Generate menu - or Command+N on a Mac - and override a method called getMatchers(). We don't need to call the parent method because its empty.

This method is... kinda beautiful: just return an array where they keys are the custom matchers you want. Except, the key is not shouldReturnZero(). Nope, the name of the matcher is that string without the 'should' or 'shouldNot' part. In other words, add returnZero set to a function with one argument called $ subject.

The Matcher Subject

Here's how this works: in the example, we call getLength(), which we know returns an integer - hopefully zero. But thanks to the magic of phpspec, we can call shouldReturnZero() on this value. When we do that, phpspec will call our function and pass the length returned from getLength() as the $ subject. Complete the matcher by saying return $ subject === 0. Our matcher function should return true if the  looks valid, false otherwise.

So... let's try this! Go spec go!
 ./vendor/bin/phpspec run
It passes! Oh, and we can automatically also use shouldNotReturnZero(): every matcher is able to handle both should and shouldNot.