Better Error Message
To make sure the matcher is really working, in Dinosaur, add a bug by changing the default length to 30.

Now re-run phpspec:
./vendor/bin/phpspec run
Two examples fail - we're working on the second example. Look at the error:

  integer:30 expected to returnZero(), but it is not.

Wow. That's... kinda bad language. phpspec is trying its best to tell us what went wrong in a way that makes sense... but it doesn't always work.

No problem: we can control that error.