Let's go fix that bug - change 30 back to zero - and re-run phpspec: ./vendor/bin/phpspec run
Nice! Oh, by the way, sometimes when you call a matcher, you may need to pass it an argument... and sometimes we don't. If we did pass an argument to the matcher function, it would be passed to our callback as the second argument. And if you pass two arguments to the matcher, these become arguments two and three... and so on - you can make the matcher as complex as you need.

Because the new matcher lives right inside the spec class, this is called an 'inline' matcher. And as nice as it is, it has one major downside: the returnZero matcher can't be re-used in any other spec classes. So next: let's create another custom matcher that can be used in our entire app.