Inline matchers are easy! I love that! But you can't reuse them across multiple spec files - that's a bummer. Fortunately, phpspec is awesome and so - of course - it does have a way to create a matcher that can be used anywhere. And... what's even better is that there are a lot of great examples to learn from.
And here's one: go to https://github.com/karriereat/phpspec-matchers - I... probably butchered that username - sorry! Anyways, this library is cool: it's just a big collection of custom matchers! Well, technically it's a phpspec extension, which means it's a 'plugin' for phpspec.

Describing a Potential Bug

Here's our next goal: our scientists are starting to grow dinosaurs for the park, but have reported a possible bug in the Dinosaur class! No problem! When you have a bug, the best thing to do is write a test for it: to describe the correct behavior so we can make sure our class has that.

In this case, someone reported that, when a dinosaur is created with a length of 15, sometimes... it shrinks! We've talked to our scientists and they say that some shrinking is ok, but a dinosaur should definitely not shrink below a length of 12. Wow, it turns out that growing a dinosaur is complex!

Let's translate this expected behavior into a new example: function it_should_not_shrink(). Set the length of the dinosaur to 15 - and notice that we do get auto-completion now that the setLength() method exists.

Then say, $ this->getlength()... but... hmm. In this pretend example, the dinosaur is allowed to shrink some but not below 12. To reflect that, let's say: ->shouldBeGreaterThan(12)

As you probably saw, that is not a real matcher. So, the tests should fail. Try them:
./vendor/bin/phpspec run
They... pass? Hmm... ah! A typo! And this proves that each example must start with it_ or its_. Try phpspec again:
 ./vendor/bin/phpspec run
There is the failure we expected.