Creating our Matcher Class
We know that we could create an inline matcher. But... I kinda want to be able to re-use this in other spec classes. To do that, we can create a matcher class. In the spec/ directory, create a Matcher directory and then a new class: BeGreaterMatcher... though this class could live anywhere. The namespace should be spec then the directory path. So spec\Matcher.

But, let's keep this class empty for now: I just want to make sure that phpspec can see our new matcher. How? Via its config! Open phpspec.yml, add a matchers: section and then, very simply, list your matcher: - spec\Matcher\BeGreaterMatcher.

That's it! It won't fully work yet of course... but let's see what happens. Run phpspec:
./vendor/bin/phpspec run
Interesting:

  Custom matcher spec\Matcher\BeGreaterMatcher does not exist.

This is basically a 'class not found' error. Copy the namespace. Yeah... that looks correct... I don't see any typos. So... what's the problem? Autoloading!

Open the composer.json file. We configured composer to be able to autoload things from the src/ directory, but we haven't configured it to be able to autoload things from the spec/ directory. phpspec does not need any autoloading to be setup to find the spec files - it handles all of that itself. But if you want to put any other, non-spec, classes in this directory - like a matcher - then we do need to set up autoloading.