Run: composer dump-autoload Autoloading the phpspec Directory
phpspec does not need any autoloading to be setup to find the spec files - it handles all of that itself. But if you want to put any other, non-spec, classes in this directory - like a matcher - then we do need to set up autoloading.
No problem: copy the autoload section, paste and change it to autoload-dev. Tell composer to expect the spec\ namespace to live in the spec/ directory.To make Composer rebuild its autoloader, run:

composer dump-autoload
Cool! Try phpspec again:
./vendor/bin/phpspec run
Much better! It does see it, and now we get:

  Custom matcher spec\Matcher\BeGreaterMatcher must implement... some Matcher interface.

Apparently all matcher classes need to implement this interface. That makes sense! Let's do that next - and finish this!