According to this friendly error, a custom matcher has one important rule: it must implement this Matcher interface. Cool! To see what these classes normally look like, let's cheat and dig into some of the core matchers themselves!
Peeking at the Core Matchers

Open vendor/phpspec/phpspec/src/PhpSpec and look in the Matcher/ directory. Say hello to all the core matchers. Check out ThrowMatcher - a matcher we'll use later to help us test exceptions. Yep! It implements that interface. Let's also look at one we're already using a lot: IdentityMatcher.

Oh - instead of implementing the Matcher interface directly, this extends a BasicMatcher class, which implements that interface, but handles a lot of the low-level work. Most of the time, you'll probably want to extend this class - it just makes life easier.

Let's go! I'll close a few other files. Then, make our BeGreaterMatcher extend BasicMatcher. I'm also going to mark the class as final. There's no reason for that - it's just a general nice practice to mark a class as final unless you intend for it to be sub-classed. Though, marking a class as final can cause issues if you need to mock it. Either way - this is not important to get this all working.

Next, go to the Code -> Generate menu or Command+N on a Mac, select 'Implement Methods' and implement all four methods that we need.

Implementing supports()

The first important method is supports(). Let's var_dump() the arguments to see what we're working with: $ name, $ subject and $ arguments. Whenever any matcher is used inside of a spec class, phpspec loops over all of the matcher classes and calls supports() to figure out which one to use.
Let's see what happens! Run phpspec:
 ./vendor/bin/phpspec run
Yep! It's dumping out every time any matcher is used - once for haveType, returnZero and down here for beGreaterThan. That's the $ name argument. The $ subject is 15 because, in our spec class, getLength() returns 15 and we're calling the matcher on that value. Finally, $ arguments is an array with just one entry: 12 - because we're passing 12 as the one argument to the shouldBeGreaterThan() matcher.