The supports() method really only needs to check to make sure that the $ name is equal to beGreaterThan: that's enough to tell phpspec that we handle that. But, a lot of times, these are written to be a bit more flexible. For example, you could use in_array() to check that $ name is one of beGreater or beGreaterThan. Then, if you want, you can even make sure the types of the subject and arguments are what we expect. We'll say that this matcher should only be used if is_numeric($ subject) and if count($ arguments) is exactly one and if that argument is also numeric and if it's Halloween after midnight. Kidding.
So, when supports() returns true, phpspec will then call the match() function on top. Our job here is to return true if everything looks good, or false otherwise - just like the inline matcher. Let's var_dump the $ subject and  one more time with a die statement - to make sure it's called like we expect.

Move over and try phpspec again:
./vendor/bin/phpspec run
Yes! We are called with 15 as the $ subject and the same $ arguments array with 12 inside.