Implementing the Failure Exception Methods
Done! If the matcher returns false, phpspec will either call getFailureException() or getNegativeFailureException() - the difference is if we're using shouldBeGreaterThan() or shouldNotBeGreaterThan(). Our job there is to return that same FailureException - I'll paste one in with a good message. This is the same type of object that we're throwing from our inline matcher.
For the other method, copy this, paste, and just tweak the language a little bit.

We now have a fully-functional custom matcher that allows us to use this new, natural, language inside of any spec file. Well... assuming it works - try it out:
./vendor/bin/phpspec run

All green! Next: it's time to talk about how phpspec fits into the entire world of testing. For example, there are functional tests, integration tests, and unit tests... and multiple tools like PHPUnit and Behat. Let's dive into this big mess of tools and buzzwords.