So there's one more little bit of theory. Well, less theory - more a strategy that you get to use with BDD or TDD and... it's really rewarding. It's called the red, green, refactor cycle. It goes like this: whenever you need to add a feature or fix a bug, you follow this simple three-step process.
First, write a test! Or, in phpspec, write an 'example' showing the behavior you want. Then, run your test to make sure it's failing - that's the 'red' part.

Second, write just enough code to get that test to pass - and no more. This is the green part of the cycle - and it's more interesting than it seems at first. The key thing with this step is that you're supposed to write just enough code to get your test to pass... not focus on writing a perfect, pretty or extensible solution.

And then, once the test is green, you're free to do step 3: refactor. My favorite thing about this cycle is that it gives you permission on step 2 to write bad or duplicated code! I love this! It lets me focus on solving the problem, without over-thinking the details. And if you do need to refactor, you can do it confidently knowing that you're not going to break anything.

Theory: Follow some of It

But ultimately... these are all just 'recommendations'. In the world of testing, there are a lot of philosophical pointers and best practices that are thrown around. At the end of the day, do whatever is best for you. Just writing any tests will make your app more robust.

In this tutorial, we're going to do things - more or less - the 'right' way - with BDD and the red-green-refactor cycle. But sometimes I do the opposite! Sometimes I write the code first and then the tests. It depends on the situation and nobody is perfect. Be pragmatic.

I also don't unit test everything - actually far from it! Earlier, we tested the getLength() and setLength() methods. Those were great examples - but that code is so simple, I would not normally test it. I unit test a method if it scares me - and then rely on integration and functional tests to cover how all the little pieces work together.

Output Formatters

Ok, as promised, after all that theory, we're going to do something fun... then keep going. We already know how to run phpspec:
./vendor/bin/phpspec run

Awesome! You can also pass a format option. This accepts a number of different values, but one of the best is pretty:

 ./vendor/bin/phpspec run --format=pretty