To activate an extension, open phpspec.yml, add extensions: and then paste the extension class name. That's it. Let's go check on the terminal... yes! It's done!
The purpose of this extension is to give us a few new formatters. One of them is called nyan.dino. Ok! Run phpspec again:
./vendor/bin/phpspec run
Hello Dino! Ok, ok - time to get back to the real work. Next: let's demystify all the magic behind phpspec by looking into the ObjectBehavior class. That's the class our spec class extends - and it is responsible for allowing us to use $ this as if we were in a Dinosaur class. Understand how that works and you'll be unstoppable!