The hardest part of phpspec for me was how weird these spec classes look. They're... complete magic! You're supposed to pretend that the  variable is a Dinosaur object... even though we're not in that class? And also... I guess that means that phpspec somehow instantiates a new Dinosaur object before it runs each example? Then, just when you get used to the weirdness of treating  like a Dinosaur object and calling real methods on it... we suddenly call a matcher method - like shouldReturn(0). What is going on?
Digging into ObjectBehavior

Let's find out. Because when I finally saw how all this worked behind the scenes, I instantly felt much more comfortable. All of this magic starts with the base ObjectBehavior class. Hold Command or Ctrl and click to open that.

Ah, ok: see that protected  property? Surprise! That is actually the underlying Dinosaur object that we're testing. Well, that's not 100% true - but imagine it is for a minute. So, at some point, phpspec instantiates a Dinosaur object and stores it on that property.

Pretty much all of the magic of this class is thanks to the __call() method. If you're not familiar with this method, that's great! It's a magic PHP method that you should probably not use - but it's perfect for phpspec. If you call a non-existent method on an object, but that class has an __call() method, instead of freaking out and throwing an error, PHP will instead execute __call() and pass it the method name and the arguments you were trying to use.

And what does ObjectBehavior do in this method? It basically calls that method on $ this->object and passes it the arguments! This is why, when we say $ this->getLength(), it works! The getLength() method does not exist on ObjectBehavior. But thanks to the __call() method, it forwards that call to the actual Dinosaur object. ObjectBehavior also has a few other methods, like __get() and __set() to forward setting properties and other stuff.

The ObjectBehavior Magic
Course: phpspec: Testing... *Designing* with a Bite Tutorial

Play
Mute
Remaining Time -1:34
Playback Rate1x
Subtitles
QualityHD
Picture In Picture
Fullscreen
Autoplay

Script
 Conversation
The hardest part of phpspec for me was how weird these spec classes look. They're... complete magic! You're supposed to pretend that the  variable is a Dinosaur object... even though we're not in that class? And also... I guess that means that phpspec somehow instantiates a new Dinosaur object before it runs each example? Then, just when you get used to the weirdness of treating  like a Dinosaur object and calling real methods on it... we suddenly call a matcher method - like shouldReturn(0). What is going on!?

Digging into ObjectBehavior
Let's find out. Because when I finally saw how all this worked behind the scenes, I instantly felt much more comfortable. All of this magic starts with the base ObjectBehavior class. Hold Command or Ctrl and click to open that.

Ah, ok: see that protected  property? Surprise! That is actually the underlying Dinosaur object that we're testing. Well, that's not 100% true - but imagine it is for a minute. So, at some point, phpspec instantiates a Dinosaur object and stores it on that property.

Pretty much all of the magic of this class is thanks to the __call() method. If you're not familiar with this method, that's great! It's a magic PHP method that you should probably not use - but it's perfect for phpspec. If you call a non-existent method on an object, but that class has an __call() method, instead of freaking out and throwing an error, PHP will instead execute __call() and pass it the method name and the arguments you were trying to use.

And what does ObjectBehavior do in this method? It basically calls that method on ->object and passes it the arguments! This is why, when we say ->getLength(), it works! The getLength() method does not exist on ObjectBehavior. But thanks to the __call() method, it forwards that call to the actual Dinosaur object. ObjectBehavior also has a few other methods, like __get() and __set() to forward setting properties and other stuff.

The Wrapped Object

Let's see what some of this looks like in the wild. Close that class and, in any of the examples, var_dump($ this). Go tests go!
./vendor/bin/phpspec run
Interesting... As we expected, $ this is really an instance of DinosaurSpec. But check out the $ object property. I lied! It is not an instance of Dinosaur! Gasp! Nope, it's some Subject object from phpspec. But inside of it is something called a WrappedObject and inside if it, yep! There is the Dinosaur object.

So, it's a bit more complex than we thought at first, but phpspec did create a Dinosaur object and set it on that object property... just wrapped inside a few other objects to help the magic.